Extracted PhysicsSystem from ServerGame
